resolving merge conflict

+++ src/main/java/com/skilldistillery/film/controllers/FilmController.java
@@ -57,9 +57,12 @@ public class FilmController {
 	@RequestMapping(path = "deleteFilm.do")
 	public ModelAndView deleteFilm(int filmId) {
 		ModelAndView mv = new ModelAndView();
-		Film film = filmDao.findFilmById(filmId);
-		filmDao.deleteFilm(film);
-		mv.setViewName("filmdeleted");
+		Film film = null;
+		film = filmDao.findFilmById(filmId);
+		if (film != null) {
+			filmDao.deleteFilm(film);
+			mv.setViewName("filmdeleted");
+		}
 		mv.addObject(film);
 		return mv;
 	}
